Added median_two_sorted_arrays.py
Bruce, here are my answers to questions 1 and 3:

1. What happens if the arrays *aren't sorted*, what is the best time complexity we can achieve in this case?

The best time complexity we can achieve is O(nlogn) since we would need to sort each array and then apply an algorithm, such as the linear time complexity one in my code, to find the median.

3. How would we push this further to O(log (m + n))?

I thought about this for a while and could not come up with a working algorithm. However, I feel that binary search with some form of partition is the key to this.